fast-backend/pyproject.toml
[project]
name = "fast-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "chromadb>=1.1.0",
    "fastapi[standard]>=0.116.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langdetect>=1.0.9",
    "lyricsgenius>=3.7.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "spotipy>=2.25.1",
]
